Phase 7: Unified Startup with automatic coordinator services
Unified peer/coordinator startup into single `tunnelmesh join` command.
Coordinators auto-start when `coordinator.enabled: true` in config.

Changes:
- Integrated replicator into coordinator server
  - Added mesh transport for HTTP-based replication
  - Created S3 adapter for replication interface
  - Auto-start replicator when clustering enabled
  - Added /api/replication/message endpoint
- Modified join command for unified startup
  - Start coordinator services before registration if enabled
  - Auto-register with localhost if no servers configured
  - Start admin/S3/NFS services after mesh join with TLS
- Removed serve command (redundant with unified join)
  - Deleted runServe() and runServeFromService() functions
  - Removed unused imports and variables
- Fixed DNS test type mismatch (ClassINET int vs uint16)

New files:
- internal/coord/replication/mesh_transport.go - HTTP transport layer
- internal/coord/replication/s3_adapter.go - S3 Store adapter

Modified files:
- internal/coord/server.go - Replicator integration
- cmd/tunnelmesh/main.go - Unified startup, serve deletion
- internal/dns/srv_test.go - Type cast fix

Startup flow:
1. If coordinator.enabled: Start coord → HTTP API → Self-register
2. Join mesh as peer (register with coordinator)
3. If coordinator: Start admin/S3/NFS with mesh TLS cert

+++ internal/dns/srv_test.go
@@ -227,7 +227,7 @@ func TestHandleSRVQuery(t *testing.T) {
 				require.True(t, ok, "record should be *dns.SRV")
 				assert.Equal(t, "_coord._tcp.tunnelmesh.", srv.Hdr.Name)
 				assert.Equal(t, dns.TypeSRV, srv.Hdr.Rrtype)
-				assert.Equal(t, dns.ClassINET, srv.Hdr.Class)
+				assert.Equal(t, uint16(dns.ClassINET), srv.Hdr.Class)
 				assert.Equal(t, uint32(300), srv.Hdr.Ttl)
 				assert.Equal(t, uint16(10), srv.Priority)
 				assert.Equal(t, uint16(10), srv.Weight)
